index_document/requirements.txt
pdfplumber
openai
python-dotenv
azure-cosmos
psutil